Address PR review feedback (#392)
- merge rceCommandUtils.ts + commandUtils.ts into one commands/commandUtils.ts
- always print 'Using config: <path>' when a config file loads; add --no-config
  to bypass auto-detection
- restore demandOption for --key/--text (per-invocation values don't belong in
  config) and trim their sections (plus remote-control/getDeviceInfo/getDevId)
  from RokuDeployConfig
- fold resolveCommandOptions into loadConfigFile via a 'section' option
  (overloaded return types); section: null flattens root-only
- move interpolateEnvVars onto util
- RokuDeployConfig: username before password
