+++ src/RokuDeploy.ts
@@ -1389,89 +1389,65 @@ export class RokuDeploy {
     /**
      * Load a `rokudeploy.json` config file (jsonc: comments and trailing commas allowed).
      * Resolves `${VAR}` environment-variable references in string values and warns about invalid
-     * device-registry entries. The library never loads config implicitly — the CLI calls this for
-     * every command; library consumers call it explicitly when they want file config.
+     * device-registry entries. With no `section`, returns the full config. With a `section`, returns
+     * the flattened options for that section: root-level values overlaid with the section's values
+     * (section wins on collision), all other sections stripped; `section: null` flattens the same
+     * way for a consumer with no section of its own. The library never loads config implicitly —
+     * the CLI calls this for every command; library consumers call it explicitly.
      */
-    public loadConfigFile(options?: LoadConfigFileOptions): RokuDeployConfig {
+    public loadConfigFile(options?: LoadConfigFileOptions & { section?: undefined }): RokuDeployConfig;
+    public loadConfigFile(options: LoadConfigFileOptions & { section: ConfigSectionName | null }): Record<string, any>;
+    public loadConfigFile(options?: LoadConfigFileOptions & { section?: ConfigSectionName | null }): RokuDeployConfig | Record<string, any> {
         const cwd = options?.cwd ?? process.cwd();
         const configPath = options?.configPath ?? path.join(cwd, 'rokudeploy.json');
 
-        if (!fsExtra.existsSync(configPath)) {
-            return {};
-        }
-        const configFileText = fsExtra.readFileSync(configPath).toString();
-        const parseErrors: ParseError[] = [];
-        const config = parseJsonc(configFileText, parseErrors, {
-            allowEmptyContent: true,
-            allowTrailingComma: true,
-            disallowComments: false
-        });
-        if (parseErrors.length > 0) {
-            throw new Error(`Error parsing "${path.resolve(configPath)}": ` + JSON.stringify(
-                parseErrors.map(x => {
-                    return {
-                        message: printParseErrorCode(x.error),
-                        offset: x.offset,
-                        length: x.length
-                    };
-                })
-            ));
-        }
-        //empty file (or a bare comment) parses to undefined
-        if (!config) {
-            return {};
-        }
-        this.interpolateEnvVars(config, configPath, []);
-        //surface broken registry entries early (but only warn: an unused bad entry shouldn't
-        //break every command — resolveDevice() still hard-fails when the entry is actually used)
-        for (const name in config.devices ?? {}) {
-            try {
-                validateDeviceConfig(config.devices[name], `Device registry entry '${name}'`);
-            } catch (e) {
-                this.logger.warn(`${configPath}: ${(e as Error).message}`);
+        let config: RokuDeployConfig = {};
+        if (fsExtra.existsSync(configPath)) {
+            const configFileText = fsExtra.readFileSync(configPath).toString();
+            const parseErrors: ParseError[] = [];
+            //empty file (or a bare comment) parses to undefined
+            config = parseJsonc(configFileText, parseErrors, {
+                allowEmptyContent: true,
+                allowTrailingComma: true,
+                disallowComments: false
+            }) ?? {};
+            if (parseErrors.length > 0) {
+                throw new Error(`Error parsing "${path.resolve(configPath)}": ` + JSON.stringify(
+                    parseErrors.map(x => {
+                        return {
+                            message: printParseErrorCode(x.error),
+                            offset: x.offset,
+                            length: x.length
+                        };
+                    })
+                ));
+            }
+            util.interpolateEnvVars(config, configPath);
+            //surface broken registry entries early (but only warn: an unused bad entry shouldn't
+            //break every command — resolveDevice() still hard-fails when the entry is actually used)
+            for (const name in config.devices ?? {}) {
+                try {
+                    validateDeviceConfig(config.devices[name], `Device registry entry '${name}'`);
+                } catch (e) {
+                    this.logger.warn(`${configPath}: ${(e as Error).message}`);
+                }
             }
         }
-        return config;
-    }
-
-    /**
-     * Flatten a loaded config down to the options for one command: root-level common values
-     * overlaid with that command's section (section wins on collision). Section keys for OTHER
-     * commands are stripped. CLI args are merged on top by the caller.
-     */
-    public resolveCommandOptions<T extends Record<string, any> = Record<string, any>>(config: RokuDeployConfig, command: ConfigSectionName): T {
+        //no section requested: hand back the whole config
+        if (options?.section === undefined) {
+            return config;
+        }
+        //flatten for one section: root values with the section overlaid, other sections stripped
         const result: Record<string, any> = {};
         for (const key in config) {
             if (!configSectionNames.includes(key as ConfigSectionName)) {
                 result[key] = config[key];
             }
         }
-        Object.assign(result, config[command]);
-        return result as T;
-    }
-
-    /**
-     * Recursively resolve `${VAR}` environment-variable references in every string value.
-     * Throws if a referenced variable is not set, naming the variable and where it was used.
-     */
-    private interpolateEnvVars(node: Record<string, any>, configPath: string, jsonPath: string[]) {
-        for (const key in node) {
-            const value = node[key];
-            if (typeof value === 'string') {
-                node[key] = value.replace(/\$\{(\w+)\}/g, (match, varName: string) => {
-                    const envValue = process.env[varName];
-                    if (envValue === undefined) {
-                        throw new Error(
-                            `Environment variable '${varName}' referenced at "${[...jsonPath, key].join('.')}" in "${path.resolve(configPath)}" is not set`
-                        );
-                    }
-                    return envValue;
-                });
-            } else if (value && typeof value === 'object') {
-                //arrays included: for-in walks indices, so entries recurse like object values
-                this.interpolateEnvVars(value, configPath, [...jsonPath, key]);
-            }
+        if (options.section !== null) {
+            Object.assign(result, config[options.section]);
         }
+        return result;
     }
 
     /**
