+++ src/commands/GetDevIdCommand.ts
@@ -3,7 +3,7 @@ import { loadCommandOptions } from './commandUtils';
 
 export class GetDevIdCommand {
     async run(args) {
-        let options = loadCommandOptions(args, 'getDevId');
+        let options = loadCommandOptions(args, null);
         const { devId } = await rokuDeploy.getDevId(options);
         console.log(devId);
     }
