+++ src/commands/GetDeviceInfoCommand.ts
@@ -3,7 +3,7 @@ import { loadCommandOptions } from './commandUtils';
 
 export class GetDeviceInfoCommand {
     async run(args) {
-        let options = loadCommandOptions(args, 'getDeviceInfo');
+        let options = loadCommandOptions(args, null);
         const outputPath = await rokuDeploy.getDeviceInfo(options);
         console.log(util.objectToTableString(outputPath));
     }
