+++ src/commands/KeyDownCommand.ts
@@ -3,7 +3,7 @@ import { loadCommandOptions } from './commandUtils';
 
 export class KeyDownCommand {
     async run(args) {
-        let options = loadCommandOptions(args, 'keyDown');
+        let options = loadCommandOptions(args, null);
         await rokuDeploy.keyDown(options);
     }
 }
