+++ src/commands/KeyPressCommand.ts
@@ -3,7 +3,7 @@ import { loadCommandOptions } from './commandUtils';
 
 export class KeyPressCommand {
     async run(args) {
-        let options = loadCommandOptions(args, 'keyPress');
+        let options = loadCommandOptions(args, null);
         await rokuDeploy.keyPress(options);
     }
 }
