+++ src/commands/KeyUpCommand.ts
@@ -3,7 +3,7 @@ import { loadCommandOptions } from './commandUtils';
 
 export class KeyUpCommand {
     async run(args) {
-        let options = loadCommandOptions(args, 'keyUp');
+        let options = loadCommandOptions(args, null);
         await rokuDeploy.keyUp(options);
     }
 }
